requirements.txt
--extra-index-url https://packages.jochem.xyz/simple
click
Sphinx
sphinx-click
coverage
flake8
python-dotenv>=0.5.1
pandas>=0.20.0rc1
jsontableschema
tableschema
coveralls
